Basic Skeleton for Sending requests from client.
The app now is able to send requests to server to sign up users. The
server is able to determine whether the ID inserted has been used before
or not. Also the java code can now validated whether the entered
password and email in the sign up are valid or not.