Parsing messages from csv files
I wrote some code to start parsing the csv by having each line in an
arraylist. Now each line(every arralylist element) will need to be
processed using tokenizers to handle strings accurately and more
efficiently. Later we should figure a way for having clickable textview
that is created dynamically according to the size of the inbox for each
user.